Changed 'maxWeightKG' to use a float for weights containing decimals.

+++ Exercises.cs
@@ -3,13 +3,13 @@
   internal class Exercises
   {
     public string Name { get; set; }
-    public int MaxWeightKG { get; set; }
+    public float MaxWeightKG { get; set; }
     public int MaxSets { get; set; }
     public int MaxReps { get; set; }
     public string TargetArea { get; set; }
     public string MuscleGroup {  get; set; }
 
-    public Exercises(string name, int maxWeightKG, string targetArea)
+    public Exercises(string name, float maxWeightKG, string targetArea)
     {
       Name = name;
       MaxWeightKG = maxWeightKG;
